mv chat history (#9725)
